requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
iniconfig==2.3.0
packaging==26.0
Pillow==12.1.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
python-telegram-bot==22.6